refactor: CategoriesFilter to not use ds-flex

- introduce LabeledButton component
- rename FilterMenu to HashtagsFilter and FilterPosts to FilterMenu
