+++ webapp/assets/_new/styles/resets.scss
@@ -19,3 +19,10 @@ h6,
 p {
   margin: 0;
 }
+
+ul,
+ol {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+}
