+++ webapp/components/FilterMenu/FilterMenu.vue
@@ -1,36 +1,46 @@
 <template>
-  <base-card class="filter-menu">
-    <h2>{{ $t('filter-menu.hashtag-search', { hashtag }) }}</h2>
+  <dropdown ref="menu" :placement="placement" :offset="offset">
     <base-button
-      icon="close"
-      circle
-      :title="this.$t('filter-menu.clearSearch')"
-      @click="clearSearch"
-    />
-  </base-card>
+      slot="default"
+      icon="filter"
+      :filled="filterActive"
+      :ghost="!filterActive"
+      slot-scope="{ toggleMenu }"
+      @click.prevent="toggleMenu()"
+    >
+      <base-icon class="dropdown-arrow" name="angle-down" />
+    </base-button>
+    <template slot="popover">
+      <ds-container>
+        <categories-filter />
+        <general-filter />
+        <language-filter />
+      </ds-container>
+    </template>
+  </dropdown>
 </template>
-
 <script>
+import Dropdown from '~/components/Dropdown'
+import { mapGetters } from 'vuex'
+import CategoriesFilter from './CategoriesFilter'
+import GeneralFilter from './GeneralFilter'
+import LanguageFilter from './LanguageFilter'
+
 export default {
-  props: {
-    hashtag: {
-      type: String,
-      required: true,
-    },
+  components: {
+    Dropdown,
+    CategoriesFilter,
+    GeneralFilter,
+    LanguageFilter,
   },
-  methods: {
-    clearSearch() {
-      this.$emit('clearSearch')
-    },
+  props: {
+    placement: { type: String },
+    offset: { type: [String, Number] },
+  },
+  computed: {
+    ...mapGetters({
+      filterActive: 'posts/isActive',
+    }),
   },
 }
 </script>
-
-<style lang="scss">
-.filter-menu.base-card {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: $space-x-small $space-base;
-}
-</style>
