+++ webapp/locales/pl.json
@@ -160,7 +160,7 @@
   "editor": {
     "placeholder": "Napisz coś inspirującego..."
   },
-  "filter-menu": {
+  "hashtags-filter": {
     "title": "Twoja bańka filtrująca"
   },
   "followButton": {
