+++ webapp/pages/index.spec.js
@@ -1,7 +1,7 @@
 import { config, shallowMount, mount } from '@vue/test-utils'
 import PostIndex from './index.vue'
 import Vuex from 'vuex'
-import FilterMenu from '~/components/FilterMenu/FilterMenu'
+import FilterCard from '~/components/FilterCard/FilterCard'
 
 const localVue = global.localVue
 
@@ -106,7 +106,7 @@ describe('PostIndex', () => {
     it('clears the search when the filter menu emits clearSearch', () => {
       mocks.$route.query.hashtag = '#samplehashtag'
       wrapper = Wrapper()
-      wrapper.find(FilterMenu).vm.$emit('clearSearch')
+      wrapper.find(FilterCard).vm.$emit('clearSearch')
       expect(wrapper.vm.hashtag).toBeNull()
     })
